+++ app/src/main/java/com/example/laza/data/CartProduct.kt
@@ -7,7 +7,7 @@ import kotlinx.parcelize.Parcelize
 data class CartProduct(
     val product: Product,
     val quantity: Int,
-    val selectedColor: Double? = null,
+    val selectedColor: Long? = null,
     val selectedSize: String? = null
 ):Parcelable{
     constructor(): this(Product(), 1, null, null)
