+++ app/src/main/java/com/example/laza/data/Product.kt
@@ -11,7 +11,7 @@ data class Product(
     val price: Float,
     val offerPercentage: Float? = null,
     val description: String? = null,
-    val colors: List<Double>? = null,
+    val colors: List<Long>? = null,
     val sizes: List<String>? = null,
     val images: List<String>
 ):Parcelable{
